Add support for skipped tests.
This was a bit tricky because it's necessary to parse the output of
a passing test to verify if a test has been skipped or not.

In addition, I implemented this slightly differently from gtest, I
output the message printed from the GTEST_SKIP macro since that seem
like the write thing to do.

Bug: 128034200

Test: New unit tests that pass, added skipped tests to some suites and
Test: verified they are listed as skipped.
Change-Id: I7a5538b9344b9883fbabeba5c4467fd5bd5446e0