Support gtest sharding options.
Add new unit tests to cover the new options.

Some updates to the unit tests.

This is slightly different from the normal gtest output. The difference
is that isolated framework prints the total number of DISABLED tests,
where gtest only lists the disabled tests it would have had in the shard.

Bug: 126376458

Test: Ran new unit tests.
Change-Id: Ib95139353302e1b3416f5f03644f01add40db099