Run unit tests on changes.
Fixed the sharding test so that it always returns results in the same
order.

Modified the parsing so that the terminated by signal message throws
away the exact name of the signal. This fixes issues where macs don't
have the same name and this fixes a problem where verify_crash is a trap
not a segfault.

Test: atest gtest_isolated_tests
Change-Id: Ia95f0736c7e5f708cb4c813a12522ab9be4c986d