Add support for --gtest_flagfile.
Remove the workaround for DeathTests since there was a bug that resulted
in it being ignored anyway. The gtest deathtest code works already because
it does a chdir to the original working directory if the arg 0 path is
relative.

Update all unit tests for slight refactor of the child_args vector.

Add new unit tests for the new option, and add one system test for it.

Bug: 141327947

Test: All new unit tests pass.
Change-Id: I6f9b7b2a76b63b93140a3a88a378fbde5af7c090