Automatically disable isolation when debugging
The process-isolation is quite annoying when one wants to debug tests
since gdb and gdbserver aren't able to track the tests into the
subprocesses. This change automatically disables the process isolation
if it looks like we are running on one of these by looking at the
'exe' of the parent process. It also adds a new flag
'--force_isolate' to disable this behavior.

Test: manual
Test: gdb $ANDROID_HOST_OUT/nativetest64/art_compiler_tests/art_compiler_tests
Test: m fdtrack_test

Change-Id: If78c1c2ba340f1d1565763757149953e9d52ce5a